+++ server/src/infrastructure/repositories/exchange-api/binance-api.repository.ts
@@ -1,9 +1,6 @@
 import { HttpService } from '@nestjs/axios';
 import { Injectable } from '@nestjs/common';
-import {
-  IExchangeApiRepository,
-  OrderTypesEnum,
-} from '../../domain/repositories/exchange-api.repository.interface';
+import { IExchangeApiRepository } from '../../../domain/repositories/exchange-api.repository.interface';
 
 const baseUrl = 'https://api.binance.com';
 
@@ -15,11 +12,7 @@ class BinanceApiRepository implements IExchangeApiRepository {
     throw new Error('Method not implemented.');
   }
 
-  createSpotOrder(
-    type: OrderTypesEnum,
-    market: string,
-    price: number,
-  ): Promise<void> {
+  createSpotOrder(): Promise<void> {
     throw new Error('Method not implemented.');
   }
